requirements.txt
# =============================================
# YOUTUBEDROP — OpenClaw Agent + Pipeline Deps
# =============================================

# --- YouTube Data & Download ---
google-api-python-client==2.136.0
google-auth==2.34.0
yt-dlp==2025.1.1
youtube-transcript-api==0.6.2

# --- Bot Integrations (Telegram + Discord) ---
python-telegram-bot[all]==21.9
discord.py==2.4.0
aiohttp==3.11.11

# --- Data & Config ---
PyYAML==6.0.2
python-dotenv==1.0.1
pydantic==2.10.4
aiosqlite==0.20.0

# --- Utilities ---
rich==13.9.4
tenacity==9.0.0
requests==2.32.3
ffmpeg-python==0.2.0
